Remove sys/sysctl.h on linux platform
This include is useless and can be removed.
This syscall has been strongly deprecated for ages and the plugin doesn't use it any more.  On old architectures, this include is benign as long as it's unused, on new ones (such as x32) it results in a build failure.
Fix bug #11477